+++ lib/index.ts
@@ -1,5 +1,6 @@
 import H264Transport from "./transports/H264Transport";
 import AACTransport from "./transports/AACTransport";
+import ONVIFMetadataTransport from "./transports/ONVIFMetadataTransport";
 import ONVIFClient from "./ONVIFClient";
 import RTSPClient from "./RTSPClient";
 import {RTPPacket, RTCPPacket} from "./util";
@@ -7,6 +8,7 @@ import {RTPPacket, RTCPPacket} from "./util";
 export {
   H264Transport,
   AACTransport,
+  ONVIFMetadataTransport,
   ONVIFClient,
   RTSPClient,
   RTPPacket,